超过4GB的文件，字节溢出的问题。
Former-commit-id: 403f18f5debd2eb8e7f0db46fe3fd27077ac9656 [formerly 88b972f5e1025300b2c81e27c67b56f55e62b0f2]
Former-commit-id: 37dc6fe1f4b6140a65975a4646401976244586bd

+++ ReadySHAREDemo/FileViewController.m
@@ -93,7 +93,7 @@ NSString *stringFromTimeInterval(NSTimeInterval t)
     NSString        *_filePath;
     NSFileHandle    *_fileHandle;
     
-    unsigned long  _downloadedBytes;
+    unsigned long  long _downloadedBytes;
     NSDate          *_timestamp;
     
     bool isAddedBySuperView;
